requirements.txt
aiodns==2.0.0
aiohttp==3.6.2
appdirs==1.4.4
arabic-reshaper==2.1.1
argon2-cffi==20.1.0
asgiref==3.2.10
astroid==2.4.2
async-generator==1.10
async-timeout==3.0.1
attrs==20.2.0
autopep8==1.5.4
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==3.2.1
boto3==1.14.53
botocore==1.17.53
CacheControl==0.12.6
cachy==0.3.0
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
cleo==0.7.6
click==7.1.2
clikit==0.4.3
cryptography==3.1
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20200713
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.1
django-allauth==0.42.0
django-discover-runner==1.0
django-extensions==3.0.9
django-import-export==2.4.0
django-quill-editor==0.1.16
django-role-permissions==3.1.0
django-storages==1.10
django-webpack-loader==0.6.0
docutils==0.15.2
email-normalize==1.0.4
entrypoints==0.3
et-xmlfile==1.0.1
filelock==3.0.12
future==0.18.2
gunicorn==20.0.4
html5lib==1.1
idna==2.10
inflection==0.5.1
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==5.5.1
jdcal==1.4.1
jedi==0.17.2
jeepney==0.4.3
Jinja2==2.11.2
jmespath==0.10.0
json5==0.9.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.6.3
jupyterlab==2.2.8
jupyterlab-pygments==0.1.2
jupyterlab-server==1.2.0
keyring==20.0.1
lazy-object-proxy==1.4.3
lockfile==0.12.2
lxml==4.5.2
mailchimp3==3.0.14
MarkupPy==1.14
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
msgpack==1.0.0
multidict==4.7.6
multitasking==0.0.9
nbclient==0.5.0
nbconvert==6.0.6
nbformat==5.0.7
nest-asyncio==1.4.1
notebook==6.1.4
nsepy==0.8
numpy==1.19.1
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.5
packaging==20.4
pandas==1.1.1
pandocfilters==1.4.2
parso==0.7.1
pastel==0.2.1
pdf2image==1.14.0
pdfkit==0.6.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
pipenv==2020.8.13
pkginfo==1.5.0.1
poetry==1.0.10
prometheus-client==0.8.0
prompt-toolkit==3.0.7
psycopg2==2.8.5
psycopg2-binary==2.8.5
ptyprocess==0.6.0
pycares==3.1.1
pycodestyle==2.6.0
pycparser==2.20
pydantic==1.6.1
Pygments==2.7.1
pylev==1.3.0
pylint==2.6.0
pyparsing==2.4.7
PyPDF2==1.26.0
pyrsistent==0.17.3
python-bidi==0.4.2
python-dateutil==2.8.1
python-decouple==3.3
python3-openid==3.2.0
pytz==2020.1
PyYAML==5.3.1
pyzmq==19.0.2
qtconsole==4.7.7
QtPy==1.9.0
-e git+https://github.com/QuantifyCapital/quantify@973fc206b9647877fadbffa6ae056474b26ebc2e#egg=quantify
reportlab==3.5.48
requests==2.24.0
requests-oauthlib==1.3.0
requests-toolbelt==0.8.0
s3transfer==0.3.3
SecretStorage==3.1.2
selenium==3.141.0
Send2Trash==1.5.0
sentry-sdk==0.17.3
shellingham==1.3.2
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
tablib==3.0.0
terminado==0.9.1
testpath==0.4.4
toml==0.10.1
tomlkit==0.5.11
tornado==6.0.4
traitlets==5.0.4
ua-parser==0.10.0
urllib3==1.25.10
user-agents==2.2.0
virtualenv==20.0.31
virtualenv-clone==0.5.4
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==5.2.0
widgetsnbextension==3.5.1
wkhtmltopdf==0.2
wrapt==1.12.1
xhtml2pdf==0.2.5
xlrd==2.0.1
xlwt==1.3.0
yarl==1.6.0
yfinance==0.1.54
